🐛 FIX: Data replication in case of undo last commit

+++ scripts/update-data-file.js
@@ -20,14 +20,20 @@ const addNewPosts = (allStagedFiles, type) => {
 		);
 
 		const existingData = require(`../${type}/data.json`);
-		const newData = newDirectories.concat(existingData);
-		const data = JSON.stringify(newData);
-
-		fs.writeFileSync(`${cwd}/${type}/data.json`, data);
-
-		console.log(
-			`Data written to ${cwd}/${type}/data.json file for ${type}`
+		const dirExist = newDirectories.every(dir =>
+			existingData.includes(dir)
 		);
+
+		if (!dirExist) {
+			const newData = newDirectories.concat(existingData);
+			const data = JSON.stringify(newData);
+
+			fs.writeFileSync(`${cwd}/${type}/data.json`, data);
+
+			console.log(
+				`Data written to ${cwd}/${type}/data.json file for ${type}`
+			);
+		}
 	}
 };
